backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Code Analysis
# ast==0.0.2  # 移除這個有問題的包，Python內建已有ast模組
bandit==1.7.5
flake8==6.1.0
black==23.11.0
mypy==1.7.1

# AI/ML
openai==1.3.7
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
python-dotenv==1.0.0
pre-commit==3.6.0 